requirements.txt
requests==2.31.0
aiohttp==3.9.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pandas>=2.2.0
newspaper3k==0.2.8
lxml
lxml_html_clean
streamlit==1.28.1
plotly==5.17.0
PyPDF2==3.0.1
schedule==1.2.0
pytz==2023.3
openai==1.3.5
asyncio==3.4.3
pinecone-client==2.2.4